refactor(tracklist-tag): restructure models and pipes under the-movie-db feature directory
